Cdr parsing and upgraded node and cell id to long

+++ trunk/mobility/src/main/java/es/tid/ps/mobility/data/MxCdrUtil.java
@@ -1,6 +1,7 @@
 package es.tid.ps.mobility.data;
 
 import com.twitter.elephantbird.mapreduce.io.ProtobufWritable;
+
 import es.tid.ps.mobility.data.BaseProtocol.Date;
 import es.tid.ps.mobility.data.BaseProtocol.Time;
 import es.tid.ps.mobility.data.MxProtocol.MxCdr;
@@ -10,7 +11,7 @@ import es.tid.ps.mobility.data.MxProtocol.MxCdr;
  * @author dmicol
  */
 public abstract class MxCdrUtil {
-    public static MxCdr create(int phone, int cell, Date date, Time time) {
+    public static MxCdr create(long phone, long cell, Date date, Time time) {
         return MxCdr.newBuilder()
                 .setPhone(phone)
                 .setCell(cell)
@@ -18,9 +19,9 @@ public abstract class MxCdrUtil {
                 .setTime(time)
                 .build();
     }
-    
-    public static ProtobufWritable createAndWrap(int phone, int cell, Date date,
-            Time time) {
+
+    public static ProtobufWritable createAndWrap(long phone, long cell,
+            Date date, Time time) {
         ProtobufWritable<MxProtocol.MxCdr> wrapper =
                 ProtobufWritable.newInstance(MxCdr.class);
         wrapper.set(create(phone, cell, date, time));
