+++ trunk/mobility/src/main/java/es/tid/ps/mobility/data/NodeBtsDayUtil.java
@@ -9,7 +9,7 @@ import es.tid.ps.mobility.data.MxProtocol.NodeBtsDay;
  * @author sortega
  */
 public abstract class NodeBtsDayUtil {
-    public static NodeBtsDay create(int node, int bts, int workday, int count) {
+    public static NodeBtsDay create(long node, int bts, int workday, int count) {
         return NodeBtsDay.newBuilder()
                 .setNode(node)
                 .setBts(bts)
@@ -18,7 +18,7 @@ public abstract class NodeBtsDayUtil {
                 .build();
     }
 
-    public static ProtobufWritable<NodeBtsDay> createAndWrap(int node, int bts,
+    public static ProtobufWritable<NodeBtsDay> createAndWrap(long node, int bts,
             int workday, int count) {
         ProtobufWritable<NodeBtsDay> wrapper =
                 ProtobufWritable.newInstance(NodeBtsDay.class);
