+++ trunk/mobility/src/main/java/es/tid/ps/mobility/jobs/MobmxFilterCellnoinfoMapper.java
@@ -1,8 +1,9 @@
 package es.tid.ps.mobility.jobs;
 
-import com.twitter.elephantbird.mapreduce.io.ProtobufWritable;
 import java.io.IOException;
-import org.apache.hadoop.io.IntWritable;
+
+import com.twitter.elephantbird.mapreduce.io.ProtobufWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.mapreduce.Mapper;
 
 import es.tid.ps.mobility.data.MxProtocol.MxCdr;
@@ -11,14 +12,14 @@ import es.tid.ps.mobility.data.MxProtocol.MxCdr;
  *
  * @author dmicol
  */
-public class MobmxFilterCellnoinfoMapper extends Mapper<IntWritable,
-        ProtobufWritable<MxCdr>, IntWritable, ProtobufWritable<MxCdr>> {
+public class MobmxFilterCellnoinfoMapper extends Mapper<LongWritable,
+        ProtobufWritable<MxCdr>, LongWritable, ProtobufWritable<MxCdr>> {
     @Override
-    public void map(IntWritable key, ProtobufWritable<MxCdr> value,
+    public void map(LongWritable key, ProtobufWritable<MxCdr> value,
             Context context) throws IOException, InterruptedException {
         MxCdr cdr = value.get();
         if (cdr.getCell() != 0) {
-            context.write(new IntWritable(cdr.getCell()), value);
+            context.write(new LongWritable(cdr.getCell()), value);
         } else {
             context.write(key, value);
         }
