+++ trunk/mobility/src/test/java/es/tid/ps/mobility/jobs/MobmxRepbtsAggbybtsTest.java
@@ -4,6 +4,7 @@ import static java.util.Arrays.asList;
 
 import com.twitter.elephantbird.mapreduce.io.ProtobufWritable;
 import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.mrunit.mapreduce.ReduceDriver;
 import org.junit.Before;
 import org.junit.Test;
@@ -17,12 +18,12 @@ import es.tid.ps.mobility.data.NodeBtsDayUtil;
  */
 public class MobmxRepbtsAggbybtsTest {
     private ReduceDriver<ProtobufWritable<NodeBtsDay>,
-        IntWritable, IntWritable, ProtobufWritable<NodeBtsDay>> driver;
+        IntWritable, LongWritable, ProtobufWritable<NodeBtsDay>> driver;
 
     @Before
     public void setUp() {
         this.driver = new ReduceDriver<ProtobufWritable<NodeBtsDay>,
-                IntWritable, IntWritable, ProtobufWritable<NodeBtsDay>>(
+                IntWritable, LongWritable, ProtobufWritable<NodeBtsDay>>(
                 new MobmxRepbtsAggbybts());
     }
 
@@ -35,7 +36,7 @@ public class MobmxRepbtsAggbybtsTest {
         this.driver
                 .withInput(NodeBtsDayUtil.createAndWrap(node, bts, workday, 101),
                            asList(new IntWritable(4), new IntWritable(5)))
-                .withOutput(new IntWritable(node),
+                .withOutput(new LongWritable(node),
                             NodeBtsDayUtil.createAndWrap(node, bts, workday, 10))
                 .runTest();
     }
